feature: match against organization UUIDs instead of IDs
