+++ .travis.yml
@@ -3,5 +3,5 @@ sudo: false
 language: ruby
 cache: bundler
 rvm:
-  - 2.6.3
+  - 2.6.6
 before_install: gem install bundler -v 2.0.2
