+++ lib/unleash-client-ruby-pipefy.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require 'unleash'
+require_relative 'unleash/strategy/org_with_uuid'
